+++ src/main/java/com/deb/dslist/services/GameService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,8 +13,8 @@ import com.deb.dslist.dto.GameDTO;
 import com.deb.dslist.dto.GameMinDTO;
 import com.deb.dslist.entities.Game;
 import com.deb.dslist.projections.GameMinProjection;
-import com.deb.dslist.repositories.GameListRepository;
 import com.deb.dslist.repositories.GameRepository;
+import com.deb.dslist.services.exceptions.DataBasesException;
 import com.deb.dslist.services.exceptions.ObjectNotFoundException;
 
 @Service
@@ -47,9 +48,15 @@ public class GameService {
 		gameRepository.save(game);
 	}
 	
-	@Transactional
+	//@Transactional(readOnly = true)
 	public void deleteGame(Long id) {
-		gameRepository.deleteById(id);
+		if(!gameRepository.existsById(id)) 
+			throw new ObjectNotFoundException(id);
+		try {
+			gameRepository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DataBasesException(e.getMessage());
+		}
 	}
 	
 }
